requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
groq>=0.4.1
python-dotenv>=1.0.0
